// src/lib/service/api-client/model/appRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * 핀플로우 프로모션: Admin API
 * 핀플로우 프로모션: Admin API 문서
 * OpenAPI spec version: 1.0
 */

/**
 * 앱 정보 저장 Request
 */
export interface AppRequest {
  /**
   * 앱 키(key)
   * @pattern ^[a-z]+$
   */
  key: string;
  /**
   * 앱 이름
   * @minLength 1
   */
  name: string;
  /**
   * postbackUrl
   * @minLength 1
   */
  postbackUrl?: string;
  /** 해시키 (입력 안하면 저장 시 자동 생성) */
  hashKey?: string;
  /** 앱 설명 */
  description?: string;
}